add http error auth
