+++ src/controllers/auth.ts
@@ -1,40 +1,12 @@
 import { Response, Request } from 'express';
 import bcrypt from 'bcryptjs';
 import { NextFunction } from 'connect';
+import createHttpError from 'http-errors';
 import db from '../models/index';
 import { generateToken } from '../utils/jwt.handle';
 import { User } from '../models/user';
-import {
-  buildUser, filterPassword, passwordHash, userByEmail,
-} from '../services/userService';
-
-const { validationResult } = require('express-validator');
-
-const registerUser = async (req: Request, res: Response) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) { return res.status(400).json({ errors: errors.array(), status: 404 }); }
-
-  const existingUser = await userByEmail(req.body.email);
-  if (existingUser) {
-    return res.status(400).json({ error: 'email already used', status: 400 });
-  }
-
-  const user = buildUser(req.body);
-  user.password = passwordHash(req.body.password);
-
-  try {
-    const newUser = await user.save();
-    const dataUser = filterPassword(newUser);
-    return res.status(200).json({ message: dataUser, status: 200 });
-  } catch (error) {
-    return res.status(404).json({ message: error, status: 404 });
-  }
-};
 
 const login = async (req: Request, res: Response, next: NextFunction) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) { return res.status(400).json({ errors: errors.array() }); }
-
   try {
     const userFromDB: User | null = await db.User.findOne({
       where: { email: req.body.email },
@@ -48,19 +20,18 @@ const login = async (req: Request, res: Response, next: NextFunction) => {
         return res.status(200).json({ message: { token }, status: 200 });
       }
     }
-
-    return res.status(403).json({ ok: false });
-  } catch (error) {
-    return next(error);
+    return next(createHttpError(401, 'Unauthorized', { contents: { ok: false } }));
+  } catch (error: Error | any) {
+    return next(createHttpError(500, error.message));
   }
 };
 
-export const authMe = async (req:Request, res:Response) => {
+export const authMe = async (req: Request, res: Response, next: NextFunction) => {
   try {
     const me = await db.User.findByPk(req.userId);
     return res.status(200).json(me);
-  } catch (error) {
-    return res.status(500).json(error);
+  } catch (error: Error | any) {
+    return next(createHttpError(500, error.message));
   }
 };
-export default { login, registerUser };
+export default { login, authMe };
